chore(sdk-coin-algo): move unit tests
- moves unit tests from bitgo & account-lib to sdk-coin-algo

Ticket: BG-48143
