--- modules/bitgo/test/v2/fixtures/coins/algo.ts
@@ -1,241 +1,0 @@
-module.exports.prebuild = function () {
-  return {
-    walletData: {
-      id: '5a78dd561c6258a907f1eeaee132f796',
-      users: [
-        {
-          user: '543c11ed356d00cb7600000b98794503',
-          permissions: [
-            'admin',
-            'view',
-            'spend',
-          ],
-        },
-      ],
-      coin: 'talgo',
-      label: 'Test Validation Wallet',
-      m: 2,
-      n: 3,
-      keys: [
-        '5a78dd56bfe424aa07aa068651b194fd',
-        '5a78dd5674a70eb4079f58797dfe2f5e',
-        '5a78dd561c6258a907f1eea9f1d079e2',
-      ],
-      tags: [
-        '5a78dd561c6258a907f1eeaee132f796',
-      ],
-      disableTransactionNotifications: false,
-      freeze: {},
-      deleted: false,
-      approvalsRequired: 1,
-      isCold: true,
-      clientFlags: [],
-      balance: 650000000,
-      confirmedBalance: 650000000,
-      spendableBalance: 650000000,
-      balanceString: '650000000',
-      confirmedBalanceString: '650000000',
-      spendableBalanceString: '650000000',
-      coinSpecific: {
-        addressVersion: 1,
-      },
-      pendingApprovals: [],
-    },
-    signedTxId: 'FYPLL4K7ARVOWBFXOCS3LAF4ZI7XZVPXVWUONYSPIRFROLDVQIOQ',
-    txData: {
-      from: 'AWSC7RL3RM72HSUW5QU4XTX3AOHY7QD3WLUZC2CAHWP6BTI5Q7IABVUXTA',
-      to: 'XMHLMNAVJIMAW2RHJXLXKKK4G3J3U6VONNO3BTAQYVDC3MHTGDP3J5OCRU',
-      amount: 1000,
-      fee: 1000,
-      firstRound: 1000,
-      lastRound: 1999,
-      genesisID: 'testnet-v38.0',
-      genesisHash: '4HkOQEL2o2bVh2P1wSpky3s6cwcEg/AAd5qlery942g=',
-      note: new Uint8Array(),
-    },
-    buildTxBase64: 'iaNhbXTNA+ijZmVlzgADsVCiZnbNA+ijZ2VurXRlc3RuZXQtdjM4LjCiZ2jEIOB5DkBC9qNm1Ydj9cEqZMt7OnMHBIPwAHeapXq8veNoomx2zQfPo3JjdsQguw62NBVKGAtqJ03XdSlcNtO6eq5rXbDMEMVGLbDzMN+jc25kxCAFpC/Fe4s/o8qW7CnLzvsDj4/Ae7LpkWhAPZ/gzR2H0KR0eXBlo3BheQ==',
-    signedAccountConsolidationBuilds: [
-      { halfSigned: { txHex:
-        'gqRtc2lng6ZzdWJzaWeTgqJwa8QgxYIFUI6hh3TlsM5dVQbzeCh+MqZGK72WRrN9Kn9cIJuhc8RA03aCdkZ9vldGalimkQvN8e4gYm90ppqAytkAuiaOvBmJK4Q+IO21KQxefKM+dZ3LVeXYj7K9ZlTs8X2aNjCAAYGicGvEIFUcn' +
-        'AbvFYmuAGFsZc+WRkrCrq07q7CfQ+jZ/XRn0UzagaJwa8QgjyzItowU95jKW/qYJcil2WAtzeEWgIcwsTTLipQqj2SjdGhyAqF2AaN0eG6Jo2FtdM4F9d0Yo2ZlZc0D6KJmds4ATFVko2dlbqx0ZXN0bmV0LXYxLjCiZ2jEIEhjtR' +
-        'iks8hOyBDyLU8QgcsPcfBZp6wg3sYvf3DlCToiomx2zgBMWUyjcmN2xCBEKgH8dLFbFC23mAChQKeOwZN9wuDi+4wUPsLKBcP4UaNzbmTEIDgbPd5+4wUXbX0bt2UbpITRF4CDMebvuEPi0Fwx38y3pHR5cGWjcGF5' } },
-      { halfSigned: { txHex:
-        'gqRtc2lng6ZzdWJzaWeTgqJwa8QgxYIFUI6hh3TlsM5dVQbzeCh+MqZGK72WRrN9Kn9cIJuhc8RAV6xhv4q7boRlzZo3XsL7/xPpvqqGUxG2sOYMTkMP8zi8TPNSWMHjT8+BXA7An1Z7jdgVOponQlJHeEjdJUK8B4GicGvEIFUcn' +
-        'AbvFYmuAGFsZc+WRkrCrq07q7CfQ+jZ/XRn0UzagaJwa8Qgvmbq2aOHbUrCz7n/bYJaGvG1VRfm3gdRAQWgSyxSIf+jdGhyAqF2AaN0eG6Jo2FtdM4F9d0Yo2ZlZc0D6KJmds4ATFVko2dlbqx0ZXN0bmV0LXYxLjCiZ2jEIEhjtR' +
-        'iks8hOyBDyLU8QgcsPcfBZp6wg3sYvf3DlCToiomx2zgBMWUyjcmN2xCBEKgH8dLFbFC23mAChQKeOwZN9wuDi+4wUPsLKBcP4UaNzbmTEIHn03U21K2bXc5ggx50tRB/pjYWbjskoKz2Jy9//VYq5pHR5cGWjcGF5' } },
-    ],
-    buildAccountConsolidation: [{
-      txHex:
-            'iaNhbXTOBfXdGKNmZWXNA+iiZnbOAExVZKNnZW6sdGVzdG5ldC12MS4womdoxCBIY7UYpLPITsgQ8i1PEIHLD3HwWaesIN7GL39w5Qk6IqJsds4ATFlMo3JjdsQgRCoB/HSxWxQtt5gAoUCnjsGTfcLg4vuMFD7CygXD+FGjc25kxCA4Gz3efuMFF219G7dlG6SE0ReAgzHm77hD4tBcMd/Mt6R0eXBlo3BheQ==',
-      txInfo:
-            { from: 'HANT3XT64MCRO3L5DO3WKG5EQTIRPAEDGHTO7OCD4LIFYMO7ZS3VJ2VRBU',
-              to: 'IQVAD7DUWFNRILNXTAAKCQFHR3AZG7OC4DRPXDAUH3BMUBOD7BISWIVYLM',
-              amount: 99999000,
-              firstRound: 5002596,
-              lastRound: 5003596,
-              genesisID: 'testnet-v1.0',
-              genesisHash: 'SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=',
-              fee: 4 },
-      feeInfo: { size: 247, fee: 1000, feeRate: 4 },
-      keys:
-            ['YWBAKUEOUGDXJZNQZZOVKBXTPAUH4MVGIYV33FSGWN6SU724ECNQLNFYKI',
-              'KUOJYBXPCWE24ADBNRS47FSGJLBK5LJ3VOYJ6Q7I3H6XIZ6RJTNAGSNCYE',
-              'R4WMRNUMCT3ZRSS37KMCLSFF3FQC3TPBC2AIOMFRGTFYVFBKR5SA5VENDM'],
-      addressVersion: 1,
-      consolidateId: '5e4cafd293a074ce39b1558b9c476097',
-      walletId: '5e4168f4403d0c5c1c3bdd15486e757f' },
-    { txHex:
-      'iaNhbXTOBfXdGKNmZWXNA+iiZnbOAExVZKNnZW6sdGVzdG5ldC12MS4womdoxCBIY7UYpLPITsgQ8i1PEIHLD3HwWaesIN7GL39w5Qk6IqJsds4ATFlMo3JjdsQgRCoB/HSxWxQtt5gAoUCnjsGTfcLg4vuMFD7CygXD+FGjc25kxCB59N1NtStm13OYIMedLUQf6Y2Fm47JKCs9icvf/1WKuaR0eXBlo3BheQ==',
-    txInfo:
-    { from: 'PH2N2TNVFNTNO44YEDDZ2LKED7UY3BM3R3ESQKZ5RHF5772VRK4RPX7LU4',
-      to: 'IQVAD7DUWFNRILNXTAAKCQFHR3AZG7OC4DRPXDAUH3BMUBOD7BISWIVYLM',
-      amount: 99999000,
-      firstRound: 5002596,
-      lastRound: 5003596,
-      genesisID: 'testnet-v1.0',
-      genesisHash: 'SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=',
-      fee: 4 },
-    feeInfo: { size: 247, fee: 1000, feeRate: 4 },
-    keys:
-        ['YWBAKUEOUGDXJZNQZZOVKBXTPAUH4MVGIYV33FSGWN6SU724ECNQLNFYKI',
-          'KUOJYBXPCWE24ADBNRS47FSGJLBK5LJ3VOYJ6Q7I3H6XIZ6RJTNAGSNCYE',
-          'XZTOVWNDQ5WUVQWPXH7W3AS2DLY3KVIX43PAOUIBAWQEWLCSEH7YZRY5ME'],
-    addressVersion: 1,
-    consolidateId: '5e4cafd293a074ce39b1558b9c476097',
-    walletId: '5e4168f4403d0c5c1c3bdd15486e757f' }]
-    ,
-    userKeychain: {
-      pub: 'UMYEHZ2NNBYX43CU37LMINSHR362FT4GFVWL6V5IHPRCJVPZ46H6CBYLYE',
-      prv: 'HEHPVOKEINTQMW4K536BZWUZWQNKBXKIAUSEWCVX3LCXDHI2LVPMV6T5L4',
-    },
-    backupKeychain: {
-      pub: 'LT7DY6JX46K2MUCLJJHEUYWKVC63FXEKCOKUSPGVTDGH3DBBK2V4SEDZWY',
-      prv: '2R4MFSNIAR4PQQGGYA6LK374X6MJEATRVNLREZ3GLAP2VYWLXDN2R4JLI4',
-    },
-    bitgoKeychain: {
-      pub: '3GFMZSPBX3XQL22IBAUF2UQFAZCOF5QFFVRRLEHFPHFLN5NY7WQJSSZRGQ',
-      // normally this wouldn't be here but we include it for signing purposes
-      prv: 'DLHE27PAI335VEQUNX7R2442SSJZUYISBL7OOR3V34AYNTH7YWI6XMGVVE',
-    },
-    signedTxBase64: 'gqRtc2lng6ZzdWJzaWeTgqJwa8QgozBD501ocX5sVN/WxDZHjv2iz4YtbL9XqDviJNX554+hc8RARmq9gs5zWBnA0TQCUBsQhzstj/v0ehzn5tPkA9WPpQ5ler5JS3SChJ9sYSDatd0Rt0TLxO6rd1MSuHiJvVRiBoGicGvEIFz+PHk355WmUEtKTkpiyqi9styKE5VJPNWYzH2MIVargaJwa8Qg2YrMyeG+7wXrSAgoXVIFBkTi9gUtYxWQ5XnKtvW4/aCjdGhyAqF2AaN0eG6Jo2FtdM0D6KNmZWXOAAOxUKJmds0D6KNnZW6tdGVzdG5ldC12MzguMKJnaMQg4HkOQEL2o2bVh2P1wSpky3s6cwcEg/AAd5qlery942iibHbNB8+jcmN2xCC7DrY0FUoYC2onTdd1KVw207p6rmtdsMwQxUYtsPMw36NzbmTEIAWkL8V7iz+jypbsKcvO+wOPj8B7sumRaEA9n+DNHYfQpHR5cGWjcGF5',
-    fullySignBase64: 'gqRtc2lng6ZzdWJzaWeTgqJwa8QgozBD501ocX5sVN/WxDZHjv2iz4YtbL9XqDviJNX554+hc8RARmq9gs5zWBnA0TQCUBsQhzstj/v0ehzn5tPkA9WPpQ5ler5JS3SChJ9sYSDatd0Rt0TLxO6rd1MSuHiJvVRiBoKicGvEIFz+PHk355WmUEtKTkpiyqi9styKE5VJPNWYzH2MIVaroXPEQJCFCAvOTZDkXG17yCZAf1dUfKy3iapk33rs8VVUM0MgcLJrTv9mJElK77pd2AWQijP8mvfslKC7EOq46ZCPQAmBonBrxCDZiszJ4b7vBetICChdUgUGROL2BS1jFZDlecq29bj9oKN0aHICoXYBo3R4bomjYW10zQPoo2ZlZc4AA7FQomZ2zQPoo2dlbq10ZXN0bmV0LXYzOC4womdoxCDgeQ5AQvajZtWHY/XBKmTLezpzBwSD8AB3mqV6vL3jaKJsds0Hz6NyY3bEILsOtjQVShgLaidN13UpXDbTunqua12wzBDFRi2w8zDfo3NuZMQgBaQvxXuLP6PKluwpy877A4+PwHuy6ZFoQD2f4M0dh9CkdHlwZaNwYXk=',
-  };
-};
-
-
-/*
- * Account lib integration test data
- *
- */
-
-export const accounts = {
-  account1: {
-    secretKey: Buffer.from('9d61b19deffd5a60ba844af492ec2cc44449c5697b326919703bac031cae7f60', 'hex'),
-    pubKey: Buffer.from('d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511a', 'hex'),
-    prvKey: 'TVQ3DHPP7VNGBOUEJL2JF3BMYRCETRLJPMZGSGLQHOWAGHFOP5QEYTZH2Q',
-    address: '25NJQAMCWEFLPVKL73J4SZAHHIHOC4XT3KTCGJNPAINGR5YHKENMEF5QTE',
-    voteKey: 'O0lMKXAaqN4h7jVJr8LHY0hvEbK62OJLJkN1soxp9Bg=',
-    selectionKey: 'wXjfzAoZhls/zrhgAn8zas/WTktvRuxeps6fkv2au+c=',
-  },
-  account2: {
-    secretKey: Buffer.from('4ccd089b28ff96da9db6c346ec114e0f5b8a319f35aba624da8cf6ed4fb8a6fb', 'hex'),
-    pubKey: Buffer.from('3d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660c', 'hex'),
-    address: 'HVABPQ7IIOEVVEVXBKTU2G36XSOJQLGPF3CJNDGAZVK7CKXUMYGA6EOE6Y',
-  },
-  account3: {
-    secretKey: Buffer.from('c5aa8df43f9f837bedb7442f31dcb7b166d38535076f094b85ce3a2e0b4458f7', 'hex'),
-    pubKey: Buffer.from('fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025', 'hex'),
-    prvKey: 'YWVI35B7T6BXX3NXIQXTDXFXWFTNHBJVA5XQSS4FZY5C4C2ELD3YWYSZLU',
-    address: '7RI43DTCDCQ2HDNEP3IAEMHQLAEBN3ITXIZQHLC55OIRKSEQQAS52OYKJE',
-  },
-  account4: {
-    secretKey: Buffer.from('9a78040f4889eac1878793757c54b85705638fad51996dd9135652ff8902d924', 'hex'),
-    pubKey: Buffer.from('e8248c0ce9cf48ef41b5648220ac84caf7ef49988503b2ef036fa16ecae7fe1c', 'hex'),
-    address: '5ASIYDHJZ5EO6QNVMSBCBLEEZL366SMYQUB3F3YDN6QW5SXH7YOLUE3ZIY',
-  },
-  default: {
-    secretKey: Buffer.from('0000000000000000000000000000000000000000000000000000000000000000', 'hex'),
-    pubKey: Buffer.from('3b6a27bcceb6a42d62a3a8d02a6f0d73653215771de243a63ac048a18b59da29', 'hex'),
-    address: 'HNVCPPGOW2SC2YVDVDICU3YNONSTEFLXDXREHJR2YBEKDC2Z3IUZSC6YGI',
-  },
-} as const;
-export const rawTx = {
-  transfer: {
-    unsigned:
-      '8ba3616d74cd2710a5636c6f7365c420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a3666565cd55f0a2667601a367656eac746573746e65742d76312e30a26768c4204863b518a4b3c84ec810f22d4f1081cb0f71f059a7ac20dec62f7f70e5093a22a26c7664a46e6f7465c4046e6f7465a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a3706179',
-    signed:
-      'gqNzaWfEQHAUb9KjDvMRU+IAxLnfKeBGA1/lrYk0F3AHTdUyBo0DI2yQfxoDzbuDTMo+UX2H8UVR9IwlaVkuinozm10z/Q6jdHhui6NhbXTNJxClY2xvc2XEIPxRzY5iGKGjjaR+0AIw8FgIFu0TujMDrF3rkRVIkIAlo2ZlZc1V8KJmdgGjZ2VurHRlc3RuZXQtdjEuMKJnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHZkpG5vdGXEBG5vdGWjcmN2xCA9QBfD6EOJWpK3CqdNG368nJgszy7ElozAzVXxKvRmDKNzbmTEINdamAGCsQq31Uv+08lkBzoO4XLz2qYjJa8CGmj3B1EapHR5cGWjcGF5',
-    // signed with account1 only
-    halfSigned:
-      '82a46d73696783a67375627369679282a2706bc420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa173c440bf584b2a9ec266c7809aaa0f848687c78d0145dc5a8967e14456d11697c72e8b9ad6b8eecbe8aea41fcf384d05a6af54e9d0928fba7d29edbdf27d9ef3c9a50081a2706bc420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a374687202a17601a374786e89a3616d74cd2710a3666565cd55f0a2667601a367656eac746573746e65742d76312e30a26768c4204863b518a4b3c84ec810f22d4f1081cb0f71f059a7ac20dec62f7f70e5093a22a26c7664a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c4205a874762a5e8f55e0a19995ae37d60a6355a26b830386eca3639ec3928660c2ca474797065a3706179',
-    // signed with account1 and account 3
-    multisig:
-      'gqRtc2lng6ZzdWJzaWeSgqJwa8Qg11qYAYKxCrfVS/7TyWQHOg7hcvPapiMlrwIaaPcHURqhc8RAv1hLKp7CZseAmqoPhIaHx40BRdxaiWfhRFbRFpfHLoua1rjuy+iupB/POE0Fpq9U6dCSj7p9Ke298n2e88mlAIKicGvEIPxRzY5iGKGjjaR+0AIw8FgIFu0TujMDrF3rkRVIkIAloXPEQN5cImZkEEELd6NLKpF/wQmWEKJ3GBjVHmRG8VvQOy8Z+8Y9dVfW/09R5hkzvcgzBk9smZbWLmG46lrGFqe3JgqjdGhyAqF2AaN0eG6Jo2FtdM0nEKNmZWXNVfCiZnYBo2dlbqx0ZXN0bmV0LXYxLjCiZ2jEIEhjtRiks8hOyBDyLU8QgcsPcfBZp6wg3sYvf3DlCToiomx2ZKNyY3bEID1AF8PoQ4lakrcKp00bfrycmCzPLsSWjMDNVfEq9GYMo3NuZMQgWodHYqXo9V4KGZla431gpjVaJrgwOG7KNjnsOShmDCykdHlwZaNwYXk=',
-  },
-  keyReg: {
-    unsigned:
-      '8ca3666565ce0004a380a2667601a367656eac746573746e65742d76312e30a26768c4204863b518a4b3c84ec810f22d4f1081cb0f71f059a7ac20dec62f7f70e5093a22a26c7664a673656c6b6579c420c178dfcc0a19865b3fceb860027f336acfd64e4b6f46ec5ea6ce9f92fd9abbe7a3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a66b6579726567a7766f746566737401a6766f74656b6409a7766f74656b6579c4203b494c29701aa8de21ee3549afc2c763486f11b2bad8e24b264375b28c69f418a7766f74656c737464',
-    signed:
-      'gqNzaWfEQHAUb9KjDvMRU+IAxLnfKeBGA1/lrYk0F3AHTdUyBo0DI2yQfxoDzbuDTMo+UX2H8UVR9IwlaVkuinozm10z/Q6jdHhui6NhbXTNJxClY2xvc2XEIPxRzY5iGKGjjaR+0AIw8FgIFu0TujMDrF3rkRVIkIAlo2ZlZc1V8KJmdgGjZ2VurHRlc3RuZXQtdjEuMKJnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHZkpG5vdGXEBG5vdGWjcmN2xCA9QBfD6EOJWpK3CqdNG368nJgszy7ElozAzVXxKvRmDKNzbmTEINdamAGCsQq31Uv+08lkBzoO4XLz2qYjJa8CGmj3B1EapHR5cGWjcGF5',
-  },
-} as const;
-
-export const explainRawTx = {
-  transfer: {
-    unsigned:
-    '8aa3616d74cd2710a3666565cd03e8a26676ce00065e82a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce0006626aa46e6f7465c40c48656c6c6f20776f726c6421a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a3706179',
-    signed:
-    '82a3736967c44002ba80a25bd0e3d6bf373fef6e839ab214036c05b0d25a366c92179eb23867e8444bb797362d9782da2a4f7e88df0757a8f1a3317640ac9611ab32a2bfd9dc0ea374786e8aa3616d74cd2710a3666565cd03e8a26676ce00065e82a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce0006626aa46e6f7465c40c48656c6c6f20776f726c6421a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a3706179',
-    halfSigned:
-    '82a46d73696783a67375627369679282a2706bc420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa173c440771ad61c57b70ef351196fb0ddb3738c6e8407171070148d4916f431fb2deffe2b8cbf027d824c8bc5c969a03735f20c6f769677cd20ae3e1a65c1c6dfb4910881a2706bc420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a374687202a17601a374786e8aa3616d74cd2710a3666565cd03e8a26676ce00065e82a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce0006626aa46e6f7465c40c48656c6c6f20776f726c6421a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c4205a874762a5e8f55e0a19995ae37d60a6355a26b830386eca3639ec3928660c2ca474797065a3706179',
-    multiSigned:
-    '82a46d73696783a67375627369679282a2706bc420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa173c440771ad61c57b70ef351196fb0ddb3738c6e8407171070148d4916f431fb2deffe2b8cbf027d824c8bc5c969a03735f20c6f769677cd20ae3e1a65c1c6dfb4910882a2706bc420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a173c440fd61b6e9da9e2ef4b78ac5c6b74f710a9c4fc847f2130a71ba2ceb37c702e0cbe82bd1a0ee6cee32f1572276bce93eb685a1a7a20e811229e2c24664fdfa140ba374687202a17601a374786e8aa3616d74cd2710a3666565cd03e8a26676ce00065e82a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce0006626aa46e6f7465c40c48656c6c6f20776f726c6421a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c4205a874762a5e8f55e0a19995ae37d60a6355a26b830386eca3639ec3928660c2ca474797065a3706179',
-    note: 'Hello world!',
-  },
-  keyreg: {
-    unsigned:
-    '8da3666565cd03e8a26676ce0006685fa367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce00066c47a46e6f7465c40c48656c6c6f20776f726c6421a673656c6b6579c420c178dfcc0a19865b3fceb860027f336acfd64e4b6f46ec5ea6ce9f92fd9abbe7a3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a66b6579726567a7766f746566737401a6766f74656b6409a7766f74656b6579c4203b494c29701aa8de21ee3549afc2c763486f11b2bad8e24b264375b28c69f418a7766f74656c737464',
-    signed:
-    '82a3736967c440996b1dd5daa598f4388cf5a926a44a3ced07fbefb411d5223dd38b2d48dd9227180e30c1eec4c46bdbc1f9ea2d4d9269ddedd5d61055df23dbacc6bf2d74b60ba374786e8da3666565cd03e8a26676ce00066898a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce00066c80a46e6f7465c40c48656c6c6f20776f726c6421a673656c6b6579c420c178dfcc0a19865b3fceb860027f336acfd64e4b6f46ec5ea6ce9f92fd9abbe7a3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a66b6579726567a7766f746566737401a6766f74656b6409a7766f74656b6579c4203b494c29701aa8de21ee3549afc2c763486f11b2bad8e24b264375b28c69f418a7766f74656c737464',
-    halfSigned:
-    '82a46d73696783a67375627369679282a2706bc420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa173c440d39f54ad6cc3e85cc46eaa453533a0968edc2342f9a8f92da753f7aff1b69bdc406ba60bfe450202d2e2b8bbc3859c0731c3712500dd8d6af3aa6ed47bed4a0d81a2706bc420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a374687202a17601a374786e8da3666565cd03e8a26676ce00066898a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce00066c80a46e6f7465c40c48656c6c6f20776f726c6421a673656c6b6579c420c178dfcc0a19865b3fceb860027f336acfd64e4b6f46ec5ea6ce9f92fd9abbe7a3736e64c4205a874762a5e8f55e0a19995ae37d60a6355a26b830386eca3639ec3928660c2ca474797065a66b6579726567a7766f746566737401a6766f74656b6409a7766f74656b6579c4203b494c29701aa8de21ee3549afc2c763486f11b2bad8e24b264375b28c69f418a7766f74656c737464',
-    multiSigned:
-    '82a46d73696783a67375627369679282a2706bc420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa173c440d39f54ad6cc3e85cc46eaa453533a0968edc2342f9a8f92da753f7aff1b69bdc406ba60bfe450202d2e2b8bbc3859c0731c3712500dd8d6af3aa6ed47bed4a0d82a2706bc420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a173c440f98e4cf8bb0844aea60a302dd071c6335f3967d86df9b5b8d3a5b96ee6f399806e832455076fed140eb5396de7b481288b9495f0b351e38db8ce8d94eb08a40aa374687202a17601a374786e8da3666565cd03e8a26676ce00066898a367656eaa73616e646e65742d7631a26768c42006bc7a0900052696118c2b0fe5dd2eb2519f88acc553d326b276aa1a5e2f2db0a26c76ce00066c80a46e6f7465c40c48656c6c6f20776f726c6421a673656c6b6579c420c178dfcc0a19865b3fceb860027f336acfd64e4b6f46ec5ea6ce9f92fd9abbe7a3736e64c4205a874762a5e8f55e0a19995ae37d60a6355a26b830386eca3639ec3928660c2ca474797065a66b6579726567a7766f746566737401a6766f74656b6409a7766f74656b6579c4203b494c29701aa8de21ee3549afc2c763486f11b2bad8e24b264375b28c69f418a7766f74656c737464',
-    note: 'Hello world!',
-  },
-  assetTransfer: {
-    unsigned:
-      '8aa461616d74cd03e8a461726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3666565cd03e8a2667601a367656ed92c53474f31474b537a7945374945504974547843427977397838466d6e724344657869392f634f554a4f69493da26768c408b5eb2d9deb7ebf5da26c760aa3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a56178666572a47861696401',
-    signed:
-      '82a3736967c440177b6d7f548c8c579618200fae3f61549cdd3fe0aba5a74cc3dc72112a5fe4aa46b209f63a364a33127886f06ffca45fcbe66cca64c990a6cbf7e8fdbace950ca374786e8aa461616d74cf8ac7230489e80000a461726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3666565cd03e8a2667601a367656eac746573746e65742d76312e30a26768c4204863b518a4b3c84ec810f22d4f1081cb0f71f059a7ac20dec62f7f70e5093a22a26c760aa3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a56178666572a47861696401',
-  },
-  USDCAssetTransfer: {
-    signed:
-      'gqNzaWfEQDWnQkjvAuBtdmhOADVtR6Ar3YJueNtWBvH0nz143kFmqC/REU7456qW1j3DR1R9DJ2JcuFNXJAGxLjm1NsMZwWjdHhuiqRhYW10AaRhcmN2xCCmatZ2ybQCH/drXFhCcsgo7/rjHdXjBkfBsY9SZdGMK6NmZWXNA+iiZnbNA+ijZ2VurHRlc3RuZXQtdjEuMKJnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHbNB9Cjc25kxCDXWpgBgrEKt9VL/tPJZAc6DuFy89qmIyWvAhpo9wdRGqR0eXBlpWF4ZmVypHhhaWTOAJ+XPQ==',
-  },
-  enableToken: '89a461726376c420e173cd0ad9e55bd835f5cd2501891cae7270cd28b7b7ef15eb2aa634b77e6164a3666565cd03e8a26676cd03e8a367656eac6d61696e6e65742d76312e30a26768c420c061c4d8fc1dbdded2d7604be4568e3f6d041987ac37bde4b620b5ab39248adfa26c76cd07d0a3736e64c420e173cd0ad9e55bd835f5cd2501891cae7270cd28b7b7ef15eb2aa634b77e6164a474797065a56178666572a478616964ce00f48c68',
-  enableTokenUSDT: 'iaRhcmN2xCDXWpgBgrEKt9VL/tPJZAc6DuFy89qmIyWvAhpo9wdRGqNmZWXNA+iiZnbNA+ijZ2VurHRlc3RuZXQtdjEuMKJnaNksU0dPMUdLU3p5RTdJRVBJdFR4Q0J5dzl4OEZtbnJDRGV4aTkvY09VSk9pST2ibHbNB9Cjc25kxCDXWpgBgrEKt9VL/tPJZAc6DuFy89qmIyWvAhpo9wdRGqR0eXBlpWF4ZmVypHhhaWTOAALA3w==',
-  enableTokenUSDC: 'iaRhcmN2xCDXWpgBgrEKt9VL/tPJZAc6DuFy89qmIyWvAhpo9wdRGqNmZWXNA+iiZnbNA+ijZ2VurHRlc3RuZXQtdjEuMKJnaNksU0dPMUdLU3p5RTdJRVBJdFR4Q0J5dzl4OEZtbnJDRGV4aTkvY09VSk9pST2ibHbNB9Cjc25kxCDXWpgBgrEKt9VL/tPJZAc6DuFy89qmIyWvAhpo9wdRGqR0eXBlpWF4ZmVypHhhaWTOAJ+XPQ==',
-  disableTokenUSDC: 'iqZhY2xvc2XEIKZq1nbJtAIf92tcWEJyyCjv+uMd1eMGR8Gxj1Jl0YwrpGFyY3bEINdamAGCsQq31Uv+08lkBzoO4XLz2qYjJa8CGmj3B1Eao2ZlZc0D6KJmds0D6KNnZW6sdGVzdG5ldC12MS4womdo2SxTR08xR0tTenlFN0lFUEl0VHhDQnl3OXg4Rm1uckNEZXhpOS9jT1VKT2lJPaJsds0H0KNzbmTEINdamAGCsQq31Uv+08lkBzoO4XLz2qYjJa8CGmj3B1EapHR5cGWlYXhmZXKkeGFpZM4An5c9',
-  disableToken: '8ba661636c6f7365c420a66ad676c9b4021ff76b5c584272c828effae31dd5e30647c1b18f5265d18c2ba461726376c420e173cd0ad9e55bd835f5cd2501891cae7270cd28b7b7ef15eb2aa634b77e6164a3666565cd03e8a26676cd03e8a367656eac6d61696e6e65742d76312e30a26768c420c061c4d8fc1dbdded2d7604be4568e3f6d041987ac37bde4b620b5ab39248adfa26c76cd07d0a46e6f7465c404a3616263a3736e64c420e173cd0ad9e55bd835f5cd2501891cae7270cd28b7b7ef15eb2aa634b77e6164a474797065a56178666572a478616964ce00f48c68',
-} as const;
-
-export const message = {
-  unsigned: '8ba3616d74cd2710a5636c6f7365c420fc51cd8e6218a1a38da47ed00230f0580816ed13ba3303ac5deb911548908025a3666565cd55f0a2667601a367656eac746573746e65742d76312e30a26768c4204863b518a4b3c84ec810f22d4f1081cb0f71f059a7ac20dec62f7f70e5093a22a26c7664a46e6f7465c4046e6f7465a3726376c4203d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660ca3736e64c420d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511aa474797065a3706179',
-  signed: '593278b78b1585e06659ff592e56695127218fea416196323049a6b8fc62a77bc10a0c0888d0dc8d9742fcc7f6a98188b2a14ee9a288edb049de88f742d59d03',
-} as const;
-
-export const networks = {
-  mainnet: {
-    genesisID: 'mainnet-v1.0',
-    genesisHash: 'wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8=',
-  },
-  testnet: {
-    genesisID: 'testnet-v1.0',
-    genesisHash: 'SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=',
-  },
-  betanet: {
-    genesisID: 'betanet-v1.0',
-    genesisHash: 'mFgazF+2uRS1tMiL9dsj01hJGySEmPN28B/TjjvpVW0=',
-  },
-} as const;
