+++ modules/sdk-coin-algo/.mocharc.yml
@@ -1,0 +1,8 @@
+require: 'ts-node/register'
+timeout: '60000'
+reporter: 'mochawesome'
+reporter-option:
+  - 'cdn=true'
+  - 'json=false'
+exit: true
+spec: ['test/unit/**/*.ts']
