+++ modules/sdk-coin-algo/src/algoToken.ts
@@ -59,7 +59,7 @@ export class AlgoToken extends Algo {
   static createTokenConstructors(): NamedCoinConstructor[] {
     const tokensCtors: NamedCoinConstructor[] = [];
     formattedAlgoTokens.forEach((config) => {
-      tokensCtors.push({ name: config.name, coinConstructor: AlgoToken.createTokenConstructor(config) });
+      tokensCtors.push({ name: config.type, coinConstructor: AlgoToken.createTokenConstructor(config) });
       if (config.alias) {
         tokensCtors.push({ name: config.alias, coinConstructor: AlgoToken.createTokenConstructor(config) });
       }
