+++ modules/sdk-coin-algo/test/fixtures/resources.ts
