+++ modules/sdk-coin-algo/tsconfig.json
@@ -7,11 +7,20 @@
     "esModuleInterop": true,
     "typeRoots": ["../../types", "./node_modules/@types", "../../node_modules/@types"]
   },
-  "include": ["src/**/*", "../../modules/bitgo/test/lib/test_bitgo"],
+  "include": ["src/**/*", "test/**/*"],
   "exclude": ["node_modules"],
   "references": [
     {
       "path": "../sdk-core"
+    },
+    {
+      "path": "../statics"
+    },
+    {
+      "path": "../utxo-lib"
+    },
+    {
+      "path": "../sdk-test"
     }
   ]
 }
